+++ src/main/java/domain/services/CustomerService.java
@@ -4,7 +4,7 @@ import domain.Customer;
 
 public class CustomerService {
 
-	public Customer createCustomer(String name, int initialBalance) {
+	public Customer createCustomer(String name, double initialBalance) {
 		return new Customer(name, initialBalance);
 	}
 }
